+++ CHANGELOG.md
@@ -5,10 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+### Deprecated
+### Removed
+### Fixed
+
+
 ## [0.12.0]
 
 ### Added
 - [issue/73] (https://github.com/podaac/hitide-backfill-tool/issues/73): Setup requester pay for forge, tig, forge-py, dmrpp
+- [issue/72] (https://github.com/podaac/hitide-backfill-tool/issues/72): Setup HiTIDE Tools as a Carpathia tool for running in the cloud
 ### Deprecated
 ### Removed
 ### Fixed